@@ -13,9 +13,9 @@ abstract class Module {
       _moduleDef[T] = Definition(Kind.FACTORY, db);
 
   /// Resolve a dependency's instance inside of a Module
-  T inject<T>() => _moduleDef[T].type == Kind.SINGLE
-      ? _moduleDef[T].instance
-      : _moduleDef[T].creator();
+  T inject<T>() => _moduleDef[T]?.type == Kind.SINGLE
+      ? _moduleDef[T]?.instance
+      : _moduleDef[T]?.creator();
 
   LinkedHashMap<Type, Definition> kDef() => _moduleDef;
 }
